Added 4 events regarding page show and hide: * beforepageshow * beforepagehide * pageshow * pagehide
All are triggered before and after the pageChange function runs.
The event.target will always be a "ui-page" div, triggered on either the showing or hiding page.

The second argument passed to the callback function contains either a prevPage or nextPage property, referencing a jQuery object containing the opposite page.