Turning on delayedClick by default. <again> Added triggers for mousedown and mouse up in the delayed case because some of hte jqm components register handlers on those events to set the component states.
Note that the focus() trigger is commented out temporarily. It seems to cause the content within the scrollview clip to scroll to the point where item being focused sits in the flow .. unfortunately this screws up the rendering because we are using CSS3 translate, so the item clicked on is out of view.

Modified sv-test-02.html so that the clone event sets the target property on the fake event.