Opt-in to disable auto-start on DOMContentLoaded.
In certain cases having the pages converted on DOMContentLoaded is not required. For example, if content is loaded from a third-party store, the page might not exist until all the data is available.
This patch allows developers to opt-out of the default initialization logic in jquery.mobile.core.js. When the content is ready and the page is rendered a simple call to $.mobile.start() will boot up jQuery Mobile and things are back to normal.