Fix orientation support to report the proper value ('portrait' or 'landscape') for devices that natively support orientation (reported in degrees).
In the presence of native orientation support, we cannot just blindly
use a ratio-based orientation measure since, on such devices, the
'orientationchange' event is typically fired BEFORE the window size is
updated.

If the browser has native orientation support, then map between device-specific
orientation values (in degrees) and a more easily used textual value
(i.e. 'portrait', 'landscape').

The mapping is based upon an initial, baseline measure of the window ratio
compared to the reported orientation which is performed once the DOM is
reported ready and the window size measure can be trusted.

If there is NOT native orientation support in the browser, fall back to a
window ratio measure initiated on window resize.