Fix for issue: https://github.com/jquery/jquery-mobile/issues/#issue/509
On initial creation of the textArea widget, we check to see if there is text and if we need to size the textArea to fit the initial text.  In order for the UE to be correct, we remove any transition effects so that the textArea just appears sized correctly.

Unfortunately, because of the way the browser runs its transition I had to use setTimeout with value of 0 in order to get the correct end user experience.  Open to suggestions on how to make this better.

Added Qunit to verify that the textArea is sized appropriately when it has default text that grows beyond the default height of 50 pixels.