Fix for issue #968: click event not passed to radio inputs
https://github.com/jquery/jquery-mobile/issues/issue/968

This fix triggers the click event for the corresponding inputs
(radio/checkbox). First the change() event is triggered, then the
click() event. This is the opposite of what Firefox does normally (first
click then change), but it is the way Chrome does it (first change then
click). I do not know if this has any further implications. This fix
also is tested on Safari on iPhone iOS 4.