mobile/event: Add swipeup and swipedown events support
Events in jQuery Mobile are handy,and very easy to use but I was quite frustrated when I discovered swipeup and swipedown events were missing.
Why should we fight with scrollstart and scrollstop events while all the code is already written but only for horizontal movements ?
I didn't update horizontalDistanceThreshold (and the like) comment to not confuse readers but if it's required, I'll submit a new patch.