Always show vertical scroll bar on .ui-mobile-viewport container
This patch prevents horizontal "jumps" in desktop browsers when a transition happens between a page that is higher than the viewport (i.e. has a scoll bar) and one that is not. If the scroll bar is not enforced it appears only after the transition finishes which causes a reflow and content "jumping" to the left.