Add a setTimeout() round updatePagePadding() for fixedToolbar
In certain cases, the updatePagePadding() function (used
to reposition elements when a page has a fixed toolbar) doesn't
reposition other page elements at the correct stage of the
page render. This can result in elements further down the <body>
being partially covered by the fixed toolbar.

To remedy this, put a setTimeout() around updatePagePadding()
so that it is applied after the rendering completes successfully.