bind to vmouseup for refresh Fixes #4756
Browsers like chrome provide an up and down arrow for range/number inputs
but they don't fire the change even until the control is blurred. Here we bind
to the mouseup and check if the value has been altered. If it has we refresh
the slider position to reflect the value change.