Add afteropen trigger to setTimeout with android fix
The select menu's (now that they depend on popup) require that the first item
is focused after they are open. The focus of the popup itself was deferred
until after the trigger of afteropen which the select menu uses to focus its
first item. As a result the focus was stolen from the select menu by the popup
to the popup container.