Transitions: Revert fbbcf9e16ed2978547f1460c66f533157714371d - was causing slow transitions on the Nexus 7. Fixes #5072.
Apparently -webkit-animation-fill-mode causes the Nexus 7 to run
animations very slowly. We didn't see this issue on phones running 4.0
or 4.1 but we're reverting this to make sure other tablets won't be
affected. IF we see it causes issues here, we're not taking chances.

We need to look into other fixes for this proposed solution.