first pass at supporting query params in the hash Fixes #5085
This commit strips any query string from the hash when referencing an internal
page. This allows users to reference exising pages but communicate information
to parts of their application and allows users to use a singple page to do
different things with the page event bindings to do so.

The information about the url is preserved under the data.options object provided
to the page lifecycle events as data.options.target which will contain the original url.