verticalswipe and horizontalswipe functions
Added veritcal swipes by creating a new function that swaps out the
variables and equations to look for vertical swipes, because vertical
layouts are just as cool. Note that this was working in a different
custom copy of jquery mobile I was working with. I though of combining
all the directions into one function but then thought it would save
little in terms of file size or computing resources, and this way is
easier for the end-user.