tests for navigate event state should work with both popstate and hashchange
This commit deserves some explanation, particularly the popstate binding
that was added. It appears at least in initial testing with Chrome, that
changing the hash _stops_ execution of the current stackframe and proceeds
to trigger a popstate event. As a result that popstate event needs to be
swallowed until we can make the replaceState modifications to the url.

In short we're keeping the even structure serialized so that the behavior
is easy to deal with. A bit of further work might store the originalEvent
for use with the synthetic version of popstate that's triggered.