Tables: Re-vamped the IE9 CSS hacks to be max-width media queries
This was needed because negating the float and width properties were
causing issues, especially on the more complex tables. Using max-width
means that we don't need to negate anything. If we can find a clean way
to negate these styles, we can simplify things a bit. All demos now
work on IE9 and are solid.