Dialog: Make sure close button inherits header theme.
When the button is added to the dialog during _create(), the header is not yet enhanced, so getInheritedTheme() works incorrectly. OTOH, since the header is not enhanced, the button need not be enhanced during create either, since it will be enhanced later when the header is itself enhanced. The button only needs to be enhanced when the option is being set on-the-fly.