Back to body theme class for count bubbles. Tweaked count bubble style.
Using the same style for the count bubbles and slider track wasn't a good idea. Because we would still need to use the ui-li-count class in the theme CSS using a fill theme class doesn't add much value either. With the new approach we add less classes with JS and it is clear to users what CSS is for the count bubble so you can delete it if you don't use them.