Navigation: Added support for authority-less URI schemes.
For URI schemes like "qrc:" (Qt resource) which do not include an
authority but only a path the method $.mobile.path.getLocation()
incorrectly returns a URL with a double slash after the scheme name.

In a Qt Webkit ThinClient application this resulted in continuously
adding up slashes in front of the path (e.g. "qrc://////index.html").