Custom select: Transfer value of "overlayTheme" option to dialog
Since 2451b0b99ad8d02e1367827741323d40f870ad91 both the "overlayTheme" option
and the "theme" option are transferred to the dialog's data-theme attribute,
resulting in two copies of the data-theme attribute being set on the dialog div.

If the "overlayTheme" option is not specified, the "theme" option is fanned out
to both the dialog's "theme" option as well as its "overlayTheme" option.

Fixes gh-6975