Autogrow: Wedge open vertical space while recalculating textarea height
A div is added after the textarea and hidden at all times except during the
textarea height calculation process. During _updateHeight() its height is set to
the current height of the textarea before the textarea height is set to 0. This
will maintain the height of the document while the textarea is being resized and
seems to successfully prevent the browser from scrolling the document -
everywhere except on Android 2.3.5, where this change makes things no worse.

We should test this change extensively.

Fixes gh-6898