IE 10 moving to background.
In IE 10 (and possibly older versions of IE) when there is a single tab open and a popup is displayed and there is another window on the same monitor, calling target.blur() can cause the browser to go to the background because the body element is the targetElement.  Ensuring the element is not the body element before calling blur() fixes this issue.