Checkboxradio: Deprecate option "mini"
This marks the occurrences of the option as deprecated, and moves processing of
the initial value of wrapperClass from _wrapper() to _setOptions(), so that
ui-mini will be applied if it's present in wrapperClass even if the mini option
is false.

Note that this requires the wrapperClass option to be processed after the mini
option in _setOptions().

Closes gh-7466
Fixes gh-7463