Fieldcontain: Reset width for notext buttons
Buttons with .ui-btn-icon-notext normally have a fixed width of 1.75em.
However when putting them as inline buttons inside a .ui-field-contain,
from a certain page width on they get a width:auto, making them use the
available width on the page instead of staying at a fixed width.
This overrides the width:auto with the 1.75em again in that situation.

Fixes gh-7526
Closes gh-7525