Widgets: Pull in new widgets from ui and remove old widgets
This commit pulls in the new jQuery UI Button, Checkboxradio, controlgroup,
and the jQuery UI Accordion. It also Updates to the newest version of
the widget factory which includes the classes option. This commit includes a
backcompat extension for switching boolean style options to classes option.
Lastely it pulls setting theme out into its own extension.